refactor(agent): split oversized fork and notice adapters
